Checkout pages moved to checkout folder and improper assertions removed
