+++ ReadMe.md
@@ -2,8 +2,6 @@
 1. Залогиниться на [странице входа](https://www.saucedemo.com/)
 2. Добавление товара Sauce Labs Fleece Jacket в корзину
 3. Открыть [страницу корзины](https://www.saucedemo.com/cart.html)
-4. Проверить количество и название товара в корзине
-5. Переход на [страницу оформления заказа](https://www.saucedemo.com/checkout-step-one.html) нажатием на кнопку "CHECKOUT"
-6. Ввод корректных данных в поля First Name, Last Name, Zip/Postal Code и переход на [страницу оформления заказа](https://www.saucedemo.com/checkout-step-one.html)
-7. Проверить информацию об оплате, доставке и итоговую цену
-8. Завершить оплату - переход на [страницу завершения оплаты](https://www.saucedemo.com/checkout-complete.html)
+4. Переход на [страницу оформления заказа](https://www.saucedemo.com/checkout-step-one.html) нажатием на кнопку "CHECKOUT"
+5. Ввод корректных данных в поля First Name, Last Name, Zip/Postal Code и переход на [страницу оформления заказа](https://www.saucedemo.com/checkout-step-two.html)
+6. Завершить оплату - переход на [страницу завершения оплаты](https://www.saucedemo.com/checkout-complete.html)
