+++ src/test/java/baseEntities/BaseTest.java
@@ -7,13 +7,11 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import steps.CheckoutSteps;
 import steps.NavigationSteps;
-import steps.ProjectSteps;
 import steps.UserSteps;
 
 public class BaseTest {
     protected WebDriver driver;
     protected UserSteps userStep;
-    protected ProjectSteps projectSteps;
     protected NavigationSteps navigationSteps;
     protected CheckoutSteps checkoutSteps;
 
@@ -23,7 +21,6 @@ public class BaseTest {
         driver.get(ReadProperties.getUrl());
 
         userStep = new UserSteps(driver);
-        projectSteps = new ProjectSteps(driver);
         navigationSteps = new NavigationSteps(driver);
         checkoutSteps = new CheckoutSteps(driver);
     }
