--- src/test/java/tests/LoginTest.java
@@ -1,29 +1,0 @@
-package tests;
-
-import baseEntities.BaseTest;
-import configuration.ReadProperties;
-import org.openqa.selenium.By;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import pages.CataloguePage;
-import pages.LoginPage;
-
-public class LoginTest extends BaseTest {
-
-    @Test
-    public void loginSuccessfulTest() {
-        Assert.assertTrue(
-                userStep.loginSuccessful(ReadProperties.username(), ReadProperties.password())
-                .isPageOpened()
-        );
-    }
-
-    @Test
-    public void addProjectTest() {
-        userStep.loginSuccessful(ReadProperties.username(), ReadProperties.password());
-        projectSteps.addProject("WP_01");
-
-        Assert.assertEquals(driver.findElement(By.className("page_title")).getText(),
-                "WP_01");
-    }
-}
