--- src/test/java/tests/LoginTest2.java
@@ -1,20 +1,0 @@
-package tests;
-
-import baseEntities.BaseTest;
-import configuration.ReadProperties;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import pages.CataloguePage;
-import pages.LoginPage;
-
-public class LoginTest2 extends BaseTest {
-
-    //@Test
-    public void loginSuccessfulTest() {
-        Assert.assertTrue(
-                userStep.loginSuccessful(ReadProperties.username(), ReadProperties.password())
-                .isPageOpened()
-        );
-    }
-
-}
